fix(tui): UserRow.profile_version is u32 after main's uint32 switch
50c810fa landed on main and switched the wire-level profile version
to uint32. The TUI's UserRow mirror still held u64 and refused to
take the new generated type. Match it here.

+++ rust/clients/qauld-ctl/src/tui/app.rs
@@ -25,7 +25,7 @@ pub struct UserRow {
     pub id: String,
     pub connectivity: String,
     pub bio: String,
-    pub profile_version: u64,
+    pub profile_version: u32,
 }
 
 #[derive(Debug, Clone)]
